WIP: RC-004-OAuth
reddit page is called and it is possible to login but afterwords it fails with a bad request.
color Scheme changed
