+++ RedditClient/Colors.xcassets/backgroundColorTwo.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "255",
-          "green" : "0",
-          "red" : "0"
+          "blue" : "137",
+          "green" : "148",
+          "red" : "247"
         }
       },
       "idiom" : "universal"
