+++ RedditClient/Colors.xcassets/backgroundTextFields.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "255",
-          "green" : "255",
-          "red" : "255"
+          "blue" : "240",
+          "green" : "241",
+          "red" : "249"
         }
       },
       "idiom" : "universal"
