+++ RedditClient/Colors.xcassets/buttonColor.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0",
-          "green" : "0",
-          "red" : "0"
+          "blue" : "24",
+          "green" : "25",
+          "red" : "39"
         }
       },
       "idiom" : "universal"
