+++ RedditClient/Login/LoginView.swift
@@ -10,8 +10,6 @@ import WebKit
 
 // swiftlint:disable multiple_closures_with_trailing_closure
 struct LoginView: View {
-    @State var username: String
-    @State var password: String
     @State var moveToListView: Bool = false
     @State var loginView: Bool = false
 
@@ -23,16 +21,16 @@ struct LoginView: View {
                         .font(.largeTitle).foregroundColor(Color.white)
                         .padding([.top, .bottom], 40)
 
+                    Spacer()
+
                     ImageView()
                         .shadow(radius: 10)
 
                     Spacer()
 
-                    UserNameTextfield(username: username)
-                    SecureTextField(password: password)
-
                     Button(action: {
-                        print("later")
+                        LoginViewModel()
+                            .doOauthReddit()
                     }) {
                             Text(R.string.localizable.login_button())
                                 .font(.headline)
@@ -43,7 +41,7 @@ struct LoginView: View {
                                 .cornerRadius(15.0)
                     }
                     .padding(.top, 30)
-                    .padding(.bottom, 30)
+                    .padding(.bottom, 60)
 
                 }.padding([.leading, .trailing], 27.5))
             .navigate(to: ListMainView(), when: $moveToListView)
@@ -58,28 +56,8 @@ struct ImageView: View {
 
         return Image(uiImage: img)
             .resizable()
-            .frame(width: 200, height: 200)
+            .frame(width: 250, height: 250)
             .padding(.bottom, 50)
-    }
-}
-
-struct UserNameTextfield: View {
-    @State var username: String
-    var body: some View {
-        TextField(R.string.localizable.username_placeholder(), text: self.$username)
-            .padding()
-            .cornerRadius(20.0)
-            .background(Color(R.color.backgroundTextFields()!))
-    }
-}
-
-struct SecureTextField: View {
-    @State var password: String
-    var body: some View {
-        SecureField(R.string.localizable.password_placeholder(), text: self.$password)
-            .padding()
-            .cornerRadius(20.0)
-            .background(Color(R.color.backgroundTextFields()!))
     }
 }
 
@@ -97,6 +75,6 @@ struct BackgroundColor: View {
 
 struct LoginView_Previews: PreviewProvider {
     static var previews: some View {
-        LoginView(username: "", password: "")
+        LoginView()
     }
 }
